Decorators common functions
