+++ .gitignore
@@ -78,7 +78,7 @@ yarn.lock
 
 # Covers JetBrains IDEs: IntelliJ, RubyMine, PhpStorm, AppCode, PyCharm, CLion, Android Studio and WebStorm
 # Reference: https://intellij-support.jetbrains.com/hc/en-us/articles/206544839
-
+.idea
 # User-specific stuff
 .idea/**/workspace.xml
 .idea/**/tasks.xml
